backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pgvector
langchain
langgraph
langchain-openai
python-dotenv
langchain-community
sqlparse  